alter texts
